create resolvers for category featured content
