+++ src/resolvers/featuredContent.ts
@@ -1,6 +1,5 @@
-import { Arg, Args, ArgsType, Authorized, Field, ID, InputType, Mutation, Query, Resolver } from 'type-graphql'
+import { Arg, Args, ArgsType, Authorized, Field, ID, InputType, Mutation, Resolver } from 'type-graphql'
 import { FeaturedVideo, getFeaturedContentDoc, VideoHero } from '../models/FeaturedContent'
-import { CategoryFeaturedVideos } from '../entities/CategoryFeaturedVideos'
 
 @InputType()
 class FeaturedVideoInput implements FeaturedVideo {
@@ -37,27 +36,6 @@ class SetCategoryFeaturedVideoArgs {
 
 @Resolver()
 export class FeaturedContentResolver {
-  @Query(() => [FeaturedVideo], { nullable: false, description: 'Get featured videos for a given video category' })
-  async categoryFeaturedVideos(@Arg('categoryId', () => ID) categoryId: string) {
-    const featuredContent = await getFeaturedContentDoc()
-    return featuredContent.featuredVideosPerCategory.get(categoryId) || []
-  }
-
-  @Query(() => [CategoryFeaturedVideos], { nullable: false, description: 'Get featured videos for all categories' })
-  async allCategoriesFeaturedVideos() {
-    const featuredContent = await getFeaturedContentDoc()
-
-    const categoriesList: CategoryFeaturedVideos[] = []
-    featuredContent.featuredVideosPerCategory.forEach((videos, categoryId) => {
-      categoriesList.push({
-        categoryId,
-        videos,
-      })
-    })
-
-    return categoriesList
-  }
-
   @Mutation(() => VideoHero, { nullable: false })
   @Authorized()
   async setVideoHero(@Arg('newVideoHero', () => VideoHeroInput) newVideoHero: VideoHeroInput) {
